% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccer.R
\name{run_soccernet}
\alias{run_soccernet}
\alias{run_clips}
\title{Run SOCcerNET/CLIPS}
\usage{
run_soccernet(df, n = 10, block_size = 8)

run_clips(df, n = 10, block_size = 20)
}
\arguments{
\item{df}{data frame of data}

\item{n}{the number of codes return (default=10)}

\item{block_size}{run the job in blocks to optimizer performance (default=8)}
}
\description{
Run SOCcerNET/CLIPS
}
